fix: align review contracts with repository guards
