+++ tests/contracts/test_review_pr_diff_annotation.py
@@ -293,7 +293,10 @@ def test_review_pr_experimental_dispatch_is_separate_and_exact() -> None:
 
     assert "STANDARD_DISPATCH_AGENTS" in step_2_9
     assert "EXPERIMENTAL_DISPATCH_AGENTS" in step_2_9
-    assert "comm -12" in step_2_9
+    assert "select_experimental_review_dispatch" in step_2_9
+    assert "standard_agent_names=STANDARD_AGENT_ALLOWLIST.split" in step_2_9
+    assert "comm -12" not in step_2_9
+    assert "EXPERIMENTAL_AGENT_ALLOWLIST" not in step_2_9
     for agent_name in (
         "pr-review-auditor-reachability",
         "pr-review-auditor-abstraction-surface",
